+++ DataLayers/Framework/ModIntegrations.cs
@@ -1,7 +1,6 @@
 using Pathoschild.Stardew.Common.Integrations.Automate;
 using Pathoschild.Stardew.Common.Integrations.BetterJunimos;
 using Pathoschild.Stardew.Common.Integrations.BetterSprinklers;
-using Pathoschild.Stardew.Common.Integrations.Cobalt;
 using Pathoschild.Stardew.Common.Integrations.LineSprinklers;
 using Pathoschild.Stardew.Common.Integrations.PelicanFiber;
 using Pathoschild.Stardew.Common.Integrations.PrismaticTools;
@@ -24,9 +23,6 @@ namespace Pathoschild.Stardew.DataLayers.Framework
 
         /// <summary>Handles access to the Better Sprinklers mod.</summary>
         public BetterSprinklersIntegration BetterSprinklers { get; }
-
-        /// <summary>Handles access to the Cobalt mod.</summary>
-        public CobaltIntegration Cobalt { get; }
 
         /// <summary>Handles access to the Line Sprinklers mod.</summary>
         public LineSprinklersIntegration LineSprinklers { get; }
@@ -53,7 +49,6 @@ namespace Pathoschild.Stardew.DataLayers.Framework
             this.Automate = new AutomateIntegration(modRegistry, monitor);
             this.BetterJunimos = new BetterJunimosIntegration(modRegistry, monitor);
             this.BetterSprinklers = new BetterSprinklersIntegration(modRegistry, monitor);
-            this.Cobalt = new CobaltIntegration(modRegistry, monitor);
             this.LineSprinklers = new LineSprinklersIntegration(modRegistry, monitor);
             this.PelicanFiber = new PelicanFiberIntegration(modRegistry, reflection, monitor);
             this.PrismaticTools = new PrismaticToolsIntegration(modRegistry, monitor);